トップページのメニューとサンプル譜面のリンクをSSRで生成
